+++ mav_msgs/include/mav_msgs/eigen_mav_msgs.h
@@ -28,6 +28,9 @@
 #include "mav_msgs/common.h"
 
 namespace mav_msgs {
+  
+/// Actuated degrees of freedom.
+enum MavActuation { DOF4 = 4, DOF6 = 6 };
 
 struct EigenAttitudeThrust {
   EigenAttitudeThrust()
@@ -107,8 +110,8 @@ struct EigenRollPitchYawrateThrust {
  */
 class EigenMavState {
  public:
-  typedef std::vector<EigenMavState, Eigen::aligned_allocator<EigenMavState>>
-      Vector;
+  typedef std::vector<EigenMavState,
+          Eigen::aligned_allocator<EigenMavState>> Vector;
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
   /// Initializes all members to zero / identity.
@@ -159,8 +162,7 @@ class EigenMavState {
 
 struct EigenTrajectoryPoint {
   typedef std::vector<EigenTrajectoryPoint,
-                      Eigen::aligned_allocator<EigenTrajectoryPoint>>
-      Vector;
+  Eigen::aligned_allocator<EigenTrajectoryPoint>> Vector;
   EigenTrajectoryPoint()
       : timestamp_ns(-1),
         time_from_start_ns(0),
@@ -171,7 +173,8 @@ struct EigenTrajectoryPoint {
         snap_W(Eigen::Vector3d::Zero()),
         orientation_W_B(Eigen::Quaterniond::Identity()),
         angular_velocity_W(Eigen::Vector3d::Zero()),
-        angular_acceleration_W(Eigen::Vector3d::Zero()) {}
+        angular_acceleration_W(Eigen::Vector3d::Zero()),
+        degrees_of_freedom(MavActuation::DOF4) {}
 
   EigenTrajectoryPoint(int64_t _time_from_start_ns,
                        const Eigen::Vector3d& _position,
@@ -181,7 +184,8 @@ struct EigenTrajectoryPoint {
                        const Eigen::Vector3d& _snap,
                        const Eigen::Quaterniond& _orientation,
                        const Eigen::Vector3d& _angular_velocity,
-                       const Eigen::Vector3d& _angular_acceleration)
+                       const Eigen::Vector3d& _angular_acceleration,
+                       const MavActuation& _degrees_of_freedom = MavActuation::DOF4)
       : time_from_start_ns(_time_from_start_ns),
         position_W(_position),
         velocity_W(_velocity),
@@ -190,7 +194,8 @@ struct EigenTrajectoryPoint {
         snap_W(_snap),
         orientation_W_B(_orientation),
         angular_velocity_W(_angular_velocity),
-        angular_acceleration_W(_angular_acceleration) {}
+        angular_acceleration_W(_angular_acceleration),
+        degrees_of_freedom(_degrees_of_freedom) {}
 
   EigenTrajectoryPoint(int64_t _time_from_start_ns,
                        const Eigen::Vector3d& _position,
@@ -199,14 +204,14 @@ struct EigenTrajectoryPoint {
                        const Eigen::Vector3d& _jerk,
                        const Eigen::Vector3d& _snap,
                        const Eigen::Quaterniond& _orientation,
-                       const Eigen::Vector3d& _angular_velocity)
+                       const Eigen::Vector3d& _angular_velocity,
+                       const MavActuation& _degrees_of_freedom = MavActuation::DOF4)
       : EigenTrajectoryPoint(_time_from_start_ns, _position, _velocity,
                              _acceleration, _jerk, _snap, _orientation,
-                             _angular_velocity, Eigen::Vector3d::Zero()) {}
+                             _angular_velocity, Eigen::Vector3d::Zero(), _degrees_of_freedom) {}
 
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  int64_t
-      timestamp_ns;  // Time since epoch, negative value = invalid timestamp.
+  int64_t timestamp_ns;  // Time since epoch, negative value = invalid timestamp.
   int64_t time_from_start_ns;
   Eigen::Vector3d position_W;
   Eigen::Vector3d velocity_W;
@@ -217,6 +222,7 @@ struct EigenTrajectoryPoint {
   Eigen::Quaterniond orientation_W_B;
   Eigen::Vector3d angular_velocity_W;
   Eigen::Vector3d angular_acceleration_W;
+  MavActuation degrees_of_freedom;
 
   // Accessors for making dealing with orientation/angular velocity easier.
   inline double getYaw() const { return yawFromQuaternion(orientation_W_B); }
@@ -252,6 +258,27 @@ struct EigenTrajectoryPoint {
   }
 };
 
+// Operator overload to transform Trajectory Points according to the Eigen
+// interfaces (uses operator* for this).
+// Has to be outside of class.
+// Example:
+// Eigen::Affine3d transform; EigenTrajectoryPoint point;
+// EigenTrajectoryPoint transformed = transform * point;
+inline EigenTrajectoryPoint operator*(const Eigen::Affine3d& lhs,
+                                      const EigenTrajectoryPoint& rhs) {
+  EigenTrajectoryPoint transformed(rhs);
+  transformed.position_W = lhs * rhs.position_W;
+  transformed.velocity_W = lhs.rotation() * rhs.velocity_W;
+  transformed.acceleration_W = lhs.rotation() * rhs.acceleration_W;
+  transformed.jerk_W = lhs.rotation() * rhs.jerk_W;
+  transformed.snap_W = lhs.rotation() * rhs.snap_W;
+  transformed.orientation_W_B = lhs.rotation() * rhs.orientation_W_B;
+  transformed.angular_velocity_W = lhs.rotation() * rhs.angular_velocity_W;
+  transformed.angular_acceleration_W =
+      lhs.rotation() * rhs.angular_acceleration_W;
+  return transformed;
+}
+
 struct EigenOdometry {
   EigenOdometry()
       : timestamp_ns(-1),
@@ -270,8 +297,7 @@ struct EigenOdometry {
         angular_velocity_B(_angular_velocity) {}
 
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  int64_t
-      timestamp_ns;  // Time since epoch, negative value = invalid timestamp.
+  int64_t timestamp_ns;  // Time since epoch, negative value = invalid timestamp.
   Eigen::Vector3d position_W;
   Eigen::Quaterniond orientation_W_B;
   Eigen::Vector3d velocity_B;  // Velocity in expressed in the Body frame!